* Use long id (instead of int id) in InstrumentConfiguration and DataEncoding as in all other 'table object'

+++ mzdb-access/src/main/java/fr/profi/mzdb/io/reader/cache/DataEncodingAsyncReader.java
@@ -48,7 +48,7 @@ public class DataEncodingAsyncReader extends AbstractDataEncodingReader {
 	 * @throws SQLiteException
 	 *             the sQ lite exception
 	 */
-	public Observable<DataEncoding> getDataEncoding(int dataEncodingId) {
+	public Observable<DataEncoding> getDataEncoding(long dataEncodingId) {
 		return mzDbReader.observeJobExecution( connection -> {
 			return this.getDataEncoding(dataEncodingId, connection);
 		});
@@ -74,7 +74,7 @@ public class DataEncodingAsyncReader extends AbstractDataEncodingReader {
 	 * @throws SQLiteException
 	 *             the sQ lite exception
 	 */
-	public Observable<Map<Integer, DataEncoding>> getDataEncodingById() {
+	public Observable<Map<Long, DataEncoding>> getDataEncodingById() {
 		return mzDbReader.observeJobExecution( connection -> {
 			return this.getDataEncodingById(connection);
 		});
